requirements.txt
# Core
apache-airflow==2.7.1
pandas
requests
pyspark==3.4.1

# Database 
dbt-postgres
apache-airflow-providers-postgres

# Quality
sqlfluff
sqlfluff-templater-dbt

# Tools
python-dotenv

# --- Framework de Tests ---
pytest